backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
pillow
pytesseract
